examples/flask-login/requirements.txt
Flask-Login
Flask-Simon
